There is a bug where the exclusive lock on the schema_migrations table is released after the first query or execution on it because the connection using it is in auto-commit mode which releases the lock after any statement completes or an execute occurs on the connection.
If this deadlock was fixed then there would be another deadlock in
Migrator#get_installed_migrations() between the connection locking the
schema_migrations table and another connection querying the table for
a list of installed migrations.  So refactor the code that allows the
connection to perform the query with to be specified.

* src/main/com/imageworks/migration/Migrator.scala
  (get_installed_migrations):
    Make this method just get a new connection and call the
    functionality that was moved from this method into...
  (get_installed_migrations_):
    ...this private method that takes a java.sql.Connection argument.