Simplify and fix ConnectionBuilder#withConnection() edge-cases.
In the previous implementation, if CommitUponReturnOrException or
CommitUponReturnOrRollbackUponException was used, then if the commit
threw an exception, it would be treated the same as the body throwing
an exception.  Hence CommitUponReturnOrException would try a second
commit and CommitUponReturnOrRollbackUponException would try to
rollback the transaction; the later being acceptable behavior.

The new implementation distinguishes between the body throwing an
exception and the following commit or rollback throwing an exception.

Change-Id: I617bb31a94373205088505b6641aa25c8f1e4431