Support MySQL which doesn't have table and column constraints.
According to the MySQL manual [1] in the ALTER TABLE clause: "The
CHECK clause is parsed but ignored by all storage engines."
Additionally, MySQL cannot parse CHECK clauses that Derby and
PostgreSQL can:

- Not being able to name the column constraint:

  CREATE TABLE person (age INT CONSTRAINT c CHECK (age >= 0));

  ERROR 1064 (42000): You have an error in your SQL syntax; check the
  manual that corresponds to your MySQL server version for the right
  syntax to use near 'CONSTRAINT c CHECK (age >= 0))' at line 1

- Not being able to list multiple CHECK clauses:

  CREATE TABLE person (age INT CHECK (age >= 0) CHECK (age < 200));

  ERROR 1064 (42000): You have an error in your SQL syntax; check the
  manual that corresponds to your MySQL server version for the right
  syntax to use near 'CHECK (age < 200))' at line 1

One solution is to still generate SQL for column check constraints by
not naming them and assuming that users will not list multiple CHECK
clauses in a single column SQL, preferring to put all the logic in a
single CHECK clause.  However, MySQL fails to parse the SQL to drop
the constraint

  mysql> ALTER TABLE person DROP CONSTRAINT chk;

  ERROR 1064 (42000): You have an error in your SQL syntax; check the
  manual that corresponds to your MySQL server version for the right
  syntax to use near 'CONSTRAINT chk' at line 1

So add to DatabaseAdapter a supportsCheckConstraints abstract val that
specifies if the database supports check constraints.

[1] https://dev.mysql.com/doc/refman/5.5/en/alter-table.html

Change-Id: I084776f90176ba0995bac8fcbbaf5555f75694ab