Extend the concept of a user to permit MySQL style accounts.
MySQL accounts consist of a user name followed by a host name,
e.g. 'foobar'@'localhost', where the host name is always required and
the user name and host name are separately quoted [1].  The quoting
requirement breaks the existing code which would quote the entire user
name, e.g. 'foobar@localhost', which would be interpreted as
'foobar@localhost'@'%'.

Goals for this commit:

1) To support Migration's grant() and revoke() methods that accept
   string user names so that a migration can be used for the currently
   supported Derby, Oracle and PostgreSQL databases along with MySQL
   in a future commit.

2) To properly quote a user name for inclusion in a SQL statement.

The only complication is that to support Derby, Oracle and PostgreSQL
with plan user names, the DatabaseAdapter needs to be able to convert
a plain string to a User object, which is provided by a UserFactory
instance.

Add these classes:

  - User, a sealed abstract class that declares a single abstract
    quoted() method which returns the user appropriately quoted for
    inclusion in a SQL statement.
  - PlainUser, a concrete User subclass used by Derby, Oracle and
    PostgreSQL.
  - UserFactory, an abstract class with an abstract nameToUser()
    method.

Add these objects:

  - User with apply(), a function to create a PlainUser.
  - PlainUserFactory with nameToUser(), a concrete UserFactory to
    create PlainUser instances.

Changes to DatabaseAdapter:

  - add an abstract val named userFactory returning a UserFactory.
  - change grantRevokeCommon(), grantSql() and revokeSql() to take
    Array[User] instead of Array[String].

Changes to Migration:

  - add two more grant() overloaded methods that take Array[User] and
    User.
  - add two more revoke() overloaded methods that take Array[User] and
    User.

[1] http://dev.mysql.com/doc/refman/5.5/en/account-names.html

Change-Id: Iec1bf8aa72cf9e49932a6788ac0803526416aa04