Wrap long lines that use a unit test migration package name.
Unfortunately, some of the changes make it look like the migration
package name is an argument to assertTrue() or assertFalse() when it
is an argument of another method, e.g. migrator.whyNotMigrated().
However, the previous formatting was already misleading in that the
"true" or "false" argument to the method (migrator.whyNotMigrated()),
already looked like it was an argument to assertTrue().

Change-Id: I2afb8987213b2b97b73b8b88aa61ccaaac377253