Grant and revoke usage privilege for PostgreSQL.
PostgreSQL requires first granting the usage privilege on a schema
before a user can utilize any privileges granted to other DB entities
such as tables or sequences.  To address this, explicitly grant the
USAGE privilege on migrate up and explicitly drop the USAGE privilege
on migrate down in the grants test migration.

Change-Id: Ifbaf940748acdbcb69bf29af0761a2a3d194aa0e